refactor: 🚚 moved `POST /admin/seed` to `POST /volunteer/seed`
improved docs, and input verification
